Access bits through uint32_fast_t union field
Although Reb_Headers are the size of a pointer (e.g. 64 bits on 64-bit
machines), it might be that 32-bit integers are processed faster,
and only 32 bits are used.  We do want to reserve the size of a full
platform pointer (uintptr_t), and alignment would give it to us even
if we didn't ask for it.  But it's possible to make the .bits union
field that accesses are currently done through a uint_fast32_t.

+++ src/include/reb-defs.h
@@ -78,7 +78,7 @@ typedef int64_t REBI64; // 64 bit integer
 typedef uint64_t REBU64; // 64 bit unsigned integer
 typedef float REBD32; // 32 bit decimal
 typedef double REBDEC; // 64 bit decimal
-typedef uintptr_t REBFLGS; // platform-pointer-size unsigned for bit flags
+typedef uint_fast32_t REBFLGS; // unsigned used for working with bit flags
 typedef uintptr_t REBLIN; // type used to store line numbers in Rebol files
 typedef uintptr_t REBTCK; // type the debug build uses for evaluator "ticks"
